% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertCoord.R
\name{autoCorPlots}
\alias{autoCorPlots}
\title{Produces auto-correlation plots of the observed state variable for lags of -+ 20 time steps for each dyad.}
\usage{
autoCorPlots(basedata, personId, dyadId, obs, time_name)
}
\arguments{
\item{basedata}{A user provided dataframe.}

\item{obs}{The name of the column in the dataframe that has the time-varying observable (e.g., the variable for which dynamics will be assessed).}

\item{time_name}{The name of the column in the dataframe that indicates sequential temporal observations.}

\item{id}{The name of the column in the dataframe that has the person-level identifier.}

\item{dyad}{The name of the column in the dataframe that has the dyad-level identifier.}
}
\description{
Produces auto-correlation plots of the observed state variable for lags of -+ 20 time steps for each dyad.
}
