% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clo.R
\name{cloCoupleOde}
\alias{cloCoupleOde}
\title{Provides the equation for a coupled oscillator model for the differential equation solver (ode) to plot}
\usage{
cloCoupleOde(t, state, parameters)
}
\description{
Provides the equation for a coupled oscillator model for the differential equation solver (ode) to plot
}
