% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clo.R
\name{cloPredTraj}
\alias{cloPredTraj}
\title{Plots the state variable's clo model predicted temporal trajectory}
\usage{
cloPredTraj(origdata, paramData, dist0name, dist1name, obsName, m1Name, m2Name)
}
\arguments{
\item{origdata}{A dataframe produced by the "dataPrep" function.}

\item{paramData}{A dataframe produced by the "indivClo" function.}

\item{dist0name}{A name for the 0-level of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the 1-level of the distinguishing variable (e.g., "Men").}

\item{obsName}{A name for the state variable (e.g., "Emotional Experience").}

\item{m1Name}{A name for the first user-specified model.}

\item{m2Name}{A name for the second user-specified model.}
}
\description{
Produces plots of the state variable's predicted temporal trajectory based on 3 versions of the clo model. The first is the "average" model, with all parameters centered on the sample averages. The other two are specified by the user by providing centering values for any desired parameters. For example, if prior analyses showed that the system variable was predicted by the damping parameter for the 0-level of the distinguishing variable, then a logical pair of models to compare would be one with that parameter centered at a low value, such as 1SD below the mean, and one with it centered at a high value, such as 1SD above the mean.
}
