% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clo.R
\name{cloResids}
\alias{cloResids}
\title{Produces histograms of the residuals from the oscillator model for each dyad.}
\usage{
cloResids(derivData, whichModel)
}
\arguments{
\item{derivData}{A dataframe that was produced with the "estDerivs" function.}

\item{whichModel}{Whether the model to be estimated is the uncoupled-oscillator ("uncoupled") or the coupled-oscillator ("coupled").}
}
\value{
The function returns histograms of the residuals from the model for each dyad (called "plots"). The plots are also written to the working directory as a pdf file called "uncoupledResid.pdf", or "coupledResid.pdf" or "inertCoordResid.pdf"
}
\description{
Produces histograms of the residuals from the oscillator model for each dyad.
}
