% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clo.R
\name{cloSysVarOut}
\alias{cloSysVarOut}
\title{Compares a base line "intercept only" model to coupled and uncoupled oscillator models for predicting the system variable from the dynamic parameters.}
\usage{
cloSysVarOut(basedata, sysVarType, dist0name, dist1name, sysVarName)
}
\arguments{
\item{basedata}{A dataframe containing the coupled oscillator parameter estimates produced by the "indivCloCouple" function.}

\item{sysVarType}{Whether the system variable is "dyad", which means both partners have the same socre, or "indiv" which means the partners can have different scores}

\item{dist0name}{A name for the level-0 of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the level-1 of the distinguishing variable (e.g., "Men").}

\item{sysVarName}{A name for the system variable being predicted (e.g., "Satisfaction").}
}
\value{
The function returns a list including: 1) the lm or lme objects containing the full results for each model(called "models"), and 2) adjusted R^2 information for each model  (called "R2"). The function also displays histograms of the residuals and plots of the predicted values against observed values for each model.
}
\description{
The dynamic parameters used in these models come from the set of clo parameter: obs_0, obs_1, d1_0, d1_1, p_obs_0, p_obs_1, p_d1_0, p_d1_1. The 3 models compared are the baseline intercept-only, the uncoupled CLO (obs_0, obs_1, d1_0, d1_1), and the full coupled CLO. The system variable can be either dyadic (sysVarType = "dyad"), where both partners have the same score (e.g., relationship length) or individual (sysVarType = "indiv"), where the partners can have different scores (e.g., age). To make it easier to read the output, the dynamic parameters are renamed as follows: obs = freq, d1 = damp, p_obs = freqCoupling, p_d1 = dampCoupling.
}
