% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clo.R
\name{cloUncoupleOde}
\alias{cloUncoupleOde}
\title{Provides the equation for an un-coupled oscillator model for the differential equation solver (ode) to plot}
\usage{
cloUncoupleOde(t, state, parameters)
}
\description{
Provides the equation for an un-coupled oscillator model for the differential equation solver (ode) to plot
}
