% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertCoord.R
\name{coordSysVarOutPlots}
\alias{coordSysVarOutPlots}
\title{Plots of the system variable predicted from the coordination parameter estimates.}
\usage{
coordSysVarOutPlots(basedata, centCoord0, centCoord1, sysVarName, dist,
  dist0name, dist1name)
}
\arguments{
\item{basedata}{A dataframe containing the inertia-coordination parameter estimates produced by the "indivInertCoord" function.}

\item{centCoord0}{A vector of low, medium and high centering values for the coordination parameter estimates for the 0-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of coordination for the 0-level partner).}

\item{centCoord1}{A vector of low, medium and high centering values for the coordination parameter estimates for the 1-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of coordination for the 1-level partner).}

\item{sysVarName}{A name for the system variable being predicted (e.g., "Satisfaction").}

\item{dist}{A number indicating which level of the distinguishing variable to produce plots for. Must be either 0 or 1.}

\item{dist0name}{A name for the 0-level of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the 1-level of the distinguishing variable (e.g., "Men").}
}
\description{
Displays bar plots for one level of the distinguishing variable, showing model predicted means and standard errors of the system variable at low and high levels of a person's own (actor) and partner's (partner) coordination parameter estimates.
}
