% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertCoord.R
\name{crossCorPlots}
\alias{crossCorPlots}
\title{Produces cross-correlation plots of the observed state variable for lags of -+ 20 time steps for each dyad.}
\usage{
crossCorPlots(basedata)
}
\arguments{
\item{basedata}{A dataframe that has the variables needed for the dataPrep function, either in the right order or as named arguments (e.g., the order and names are: id, dyad, obs, sysVar, dist, time_name). See documentation for the dataPrep function for more detail.}
}
\description{
Produces cross-correlation plots of the observed state variable for lags of -+ 20 time steps for each dyad.
}
