% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clo.R
\name{dampSysVarOutPlots}
\alias{dampSysVarOutPlots}
\title{Plots of the system variable predicted from the "damp-only" version of the clo model.}
\usage{
dampSysVarOutPlots(basedata, centFreqSelf0, centDampSelf0, centFreqPart0,
  centDampPart0, centFreqSelf1, centDampSelf1, centFreqPart1, centDampPart1,
  sysVarName, dist, dist0name, dist1name)
}
\arguments{
\item{basedata}{A dataframe containing the clo parameter estimates produced by the "indivClo" function.}

\item{centFreqSelf0}{A vector of low, medium and high centering values for the frequency parameter estimates for the 0-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of frequency for the 0-level partner).}

\item{centDampSelf0}{A vector of low, medium and high centering values for the damping/amplification parameter estimates for the 0-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of damping/amplification for the 0-level partner).}

\item{centFreqPart0}{A vector of low, medium and high centering values for the partner influence on frequency parameter estimates for the 0-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of partner influence on frequency for the 0-level partner).}

\item{centDampPart0}{A vector of low, medium and high centering values for the partner influence on damping/amplification parameter estimates for the 0-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of partner influence on damping/amplification for the 0-level partner).}

\item{centFreqSelf1}{A vector of low, medium and high centering values for the frequency parameter estimates for the 1-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of frequency for the 1-level partner).}

\item{centDampSelf1}{A vector of low, medium and high centering values for the damping/amplification parameter estimates for the 1-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of damping/amplification for the 1-level partner).}

\item{centFreqPart1}{A vector of low, medium and high centering values for the partner influence on frequency parameter estimates for the 1-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of partner influence on frequency for the 1-level partner).}

\item{centDampPart1}{A vector of low, medium and high centering values for the partner influence on damping/amplification parameter estimates for the 1-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of partner influence on damping/amplification for the 1-level partner).}

\item{sysVarName}{A name for the system variable being predicted (e.g., "Satisfaction").}

\item{dist}{A number indicating which level of the distinguishing variable to produce plots for. Must be either 0 or 1.}

\item{dist0name}{A name for the 0-level of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the 1-level of the distinguishing variable (e.g., "Men").}
}
\description{
Displays bar plots for one level of the distinguishing variable, showing model predicted means and standard errors of the system variable at low and high levels of a person's own damping/amplification parameter estimates and low and high levels of the partner's influence on their damping/amplification parameter estimates.
}
