% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{dataPrep}
\alias{dataPrep}
\title{Reformat a user-provided dataframe in a generic form appropriate for \emph{rties} modeling}
\usage{
dataPrep(basedata, personId, dyadId, obs, dist, time_name,
  time_lag = NULL, robustScale = FALSE)
}
\arguments{
\item{basedata}{A user-provided dataframe that includes all variables needed for an rties analysis, including potential system variables, control variables, etc.}

\item{personId}{The name of the column in the dataframe that has the person-level identifier.}

\item{dyadId}{The name of the column in the dataframe that has the dyad-level identifier.}

\item{obs}{The name of the column in the dataframe that has the time-varying observable (e.g., the variable for which dynamics will be assessed).}

\item{dist}{The name of the column in the dataframe that has a variable that distinguishes the partners (e.g., sex, mother/daughter, etc) that is numeric and scored 0/1.}

\item{time_name}{The name of the column in the dataframe that indicates sequential temporal observations.}

\item{time_lag}{An optional argument for the number of lags for the lagged observable.}

\item{robustScale}{An optional argument to perform robust scaling of the de-trended observed state variable, one person at a time, using the DescTools package}
}
\value{
The function returns a dataframe that has all the variables needed for modeling system dynamics, each renamed to a generic variable name, which are:
\itemize{
\item id = person id
\item dyad = dyad id
\item obs = observed state variable
\item dist1 = 0/1 variable where the 1's indicate the 1's in the original distinguishing variable
\item time = the variable indicating temporal sequence
\item dist0 = 0/1 variable where the 1's indicate the 0's in the original distinguishing variable
\item obs_deTrend = the observed state variable with each person's linear trend removed
\item p_ = all the same variables, but for a person's partner rather than themselves
}
}
\description{
In the dataframe, the partners within each dyad must have the same number of observations (e.g. rows of data), although those can include rows that have missing values (NAs). Each dyad, however, can have it's own unique number of observations. See the "overview_data_prep" vignette for complete details on the necessary format for the dataframe.
}
