% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{dataRobScale}
\alias{dataRobScale}
\title{Apply robust scaling from the DescTools package one person at a time to the detrended observed variable (obs_deTrend).}
\usage{
dataRobScale(basedata)
}
\description{
Apply robust scaling from the DescTools package one person at a time to the detrended observed variable (obs_deTrend).
}
