% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysVarPlots.R
\name{dyadic}
\alias{dyadic}
\title{Produces plots for sysVarIn when sysVar is dyadic.}
\usage{
dyadic(basedata, sysVar_name)
}
\arguments{
\item{basedata}{A dataframe created internally by the "sysVarInPlots" function.}

\item{sysVar_name}{The name of the variable in the dataframe that contains the system variable.}
}
\value{
A plot with the profiles on the y-axis and the system variable on the x-axis
}
\description{
Produces plots for sysVarIn when sysVar is dyadic.
}
