% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{histAll}
\alias{histAll}
\title{Histograms for all numeric variables in a dataframe.}
\usage{
histAll(basedata)
}
\arguments{
\item{basedata}{A user-provided dataframe.}
}
\description{
Useful for checking distributions of potential system variables to assess normality
}
