% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysVarPlots.R
\name{indiv2profilesCont}
\alias{indiv2profilesCont}
\title{Produces plots for sysVarIn when sysVar is continuous and there are 2 profiles}
\usage{
indiv2profilesCont(testModel, sysVar0name, sysVar1name)
}
\arguments{
\item{testModel}{The model object created by sysVarIn for the interaction model (e.g., sysVarInteract)}

\item{sysVar0name}{The name created by sysVarInPlots referring to the system variable for partner-0.}

\item{sysVar1name}{The name created by sysVarInPlots referring to the system variable for partner-1.}
}
\value{
A plot produced by the interactions package.
}
\description{
Produces plots for sysVarIn when sysVar is continuous and there are 2 profiles
}
