% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysVarPlots.R
\name{indiv4profilesCat}
\alias{indiv4profilesCat}
\title{Produces plots for sysVarIn when sysVar is categorical and there are 4 profiles}
\usage{
indiv4profilesCat(basedata, testModel, sysVar0name, sysVar1name)
}
\arguments{
\item{basedata}{A dataframe created internally by the "sysVarInPlots" function.}

\item{testModel}{The model object created by sysVarIn for the interaction model (e.g., sysVarInteract)}

\item{sysVar0name}{The name created by sysVarInPlots referring to the system variable for partner-0.}

\item{sysVar1name}{The name created by sysVarInPlots referring to the system variable for partner-1.}
}
\value{
A list of 4 plots showing the simple slopes for each of the profiles.
}
\description{
Produces plots for sysVarIn when sysVar is categorical and there are 4 profiles
}
