% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysVarPlots.R
\name{indiv4profilesCont}
\alias{indiv4profilesCont}
\title{Produces plots for sysVarIn when sysVar is continuous and there are 4 profiles}
\usage{
indiv4profilesCont(prob, sysVar0name, sysVar1name)
}
\arguments{
\item{prob}{A dataframe created internally by the "sysVarInPlots" function.}

\item{sysVar0name}{The name created by sysVarInPlots referring to the system variable for partner-0.}

\item{sysVar1name}{The name created by sysVarInPlots referring to the system variable for partner-1.}
}
\value{
A list of 4 plots showing the simple slopes for each of the profiles.
}
\description{
Produces plots for sysVarIn when sysVar is continuous and there are 4 profiles
}
