% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clo.R
\name{indivCloUncouple}
\alias{indivCloUncouple}
\title{Estimates an "un-coupled oscillator" model (e.g., only self frequency and damping with no coupling) for each dyad.}
\usage{
indivCloUncouple(basedata, idConvention, dist0name, dist1name, obsName)
}
\arguments{
\item{basedata}{A dataframe that was produced with the "estDerivs" function.}

\item{idConvention}{The number that was added to the dist0 partner to get the ID number for the dist1 partner.}

\item{dist0name}{A name for the level-0 of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the level-1 of the distinguishing variable (e.g., "Men").}

\item{obsName}{A name for the observed state variables being plotted (e.g., "Emotional Experience").}
}
\value{
The function returns a list including: 1) the adjusted R^2 for the model for each dyad (called "r2"), 2) the parameter estimates for the model for each dyad (called "paramData", for use in either predicting, or being predicted by, the system variable), and 3) plots of the predicted values against the observed values for each dyad (called "plots"). The plots are also written to the working directory as a pdf file called "cloPlotsUncouple.pdf"
}
\description{
The second derivatives of the observed state variables (with linear trends removed) are predicted from each person's own observed state variable (again with linear trends removed), as well as each person's own first derivatives of the observed state variables (again with linear trends removed).
}
