% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertCoord.R
\name{indivCoord}
\alias{indivCoord}
\title{Estimates a "coordination only" model for each dyad.}
\usage{
indivCoord(basedata, dist0name, dist1name, obsName)
}
\arguments{
\item{basedata}{A dataframe that was produced with the "dataPrep" function.}

\item{dist0name}{A name for the 0-level of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the 1-level of the distinguishing variable (e.g., "Men").}

\item{obsName}{A name for the observed state variables being plotted (e.g., "Emotional Experience").}
}
\value{
The function returns a list including: 1) the adjusted R^2 for the model for each dyad (called "r2", 2) the parameter estimates for the model for each dyad (called "paramData", for use in either predicting, or being predicted by, the system variable), and 3) plots of the predicted values against the observed values for each dyad (called "plots").
}
\description{
The observed state variables (with linear trends removed) are predicted from each person's intercept and each person's partner's concurrent state variable (again with linear trends removed).
}
