% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertCoord.R
\name{indivInertCoordPlots}
\alias{indivInertCoordPlots}
\title{Produces plots of versions of the inertia-coordination model-predicted trajectories overlaid on raw data for each dyad.}
\usage{
indivInertCoordPlots(basedata, whichModel, dist0name, dist1name, obsName,
  minMax = NULL)
}
\arguments{
\item{basedata}{A dataframe that was produced with the "dataPrep" function.}

\item{whichModel}{Whether the model to be estimated is the inertia only model ("inert"), the coordination only model ("coord"), or the full inertia-coordination model ("inertCoord").}

\item{dist0name}{A name for the level-0 of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the level-1 of the distinguishing variable (e.g., "Men").}

\item{obsName}{A name for the observed state variables being plotted (e.g., "Emotional Experience").}

\item{minMax}{An optional vector with desired minimum and maximum quantiles to be used for setting the y-axis range on the plots, e.g., minMax <- c(.1, .9) would set the y-axis limits to the 10th and 90th percentiles of the observed state variables. If not provided, the default is to use the minimum and maximum observed values of the state variables.}
}
\value{
The function returns plots of the predicted values against the observed values for each dyad (called "plots"). The plots are also written to the working directory as a pdf file called "inertPlots.pdf", or "coordPlots.pdf" or "inertCoordPlots.pdf"
}
\description{
The observed state variables (with linear trends removed) are predicted from one of the 3 versions of the inertia-coordination model (inertia only, "inert"; coordination only, "coord"; full inertia-coordination, "inertCoord") for each dyad individually. The predicted trajectories are plotted overlaid on the observed trajectories.
}
