% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{indivLinear}
\alias{indivLinear}
\title{Estimates dyadic linear growth models for each dyad.}
\usage{
indivLinear(basedata, dist0name, dist1name, obsName)
}
\arguments{
\item{basedata}{A dataframe that was produced with the "dataPrep" function.}

\item{dist0name}{A name for the 0-level of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the 1-level of the distinguishing variable (e.g., "Men").}

\item{obsName}{A name for the observed state variables being plotted (e.g., "Emotional Experience").}
}
\value{
The function returns a list including: 1) the adjusted R^2 for the model for each dyad (called "r2"), 2) the parameter estimates for the model for each dyad (called "paramData", for use in either predicting, or being predicted by, the system variable), and 3) plots of the predicted values against the observed values for each dyad (called "plots"). The plots are also written to the working directory as a pdf file called "linearPlots.pdf"
}
\description{
The observed state variables are predicted from each person's intercept and each person's linear slope over time. Time is centered at the last observation, so intercepts refer to predicted values for each person at the end of the observed time period.
}
