% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertCoord.R
\name{inertCoordPredTraj}
\alias{inertCoordPredTraj}
\title{Plots the bivariate state variables' model-predicted temporal trajectories for each latent profile of inertia-coordination parameters.}
\usage{
inertCoordPredTraj(origData, lpaData, lpaParams, n_profiles,
  time_length = NULL, dist0name = NULL, dist1name = NULL,
  obsName = NULL, minMax = NULL, printPlots = T, numPlots = NULL)
}
\arguments{
\item{origData}{A dataframe that was produced with the "dataPrep" function.}

\item{lpaData}{A dataframe containing the LPA profile memberships.}

\item{lpaParams}{A matrix containing the inertia-coordination parameter estimates associated with each of the latent profiles.}

\item{n_profiles}{The number of latent profiles.}

\item{time_length}{An optional value specifying how many time points to plot across. Default is 20.}

\item{dist0name}{An optional name for the level-0 of the distinguishing variable (e.g., "Women"). Default is dist0.}

\item{dist1name}{An optional name for the level-1 of the distinguishing variable (e.g., "Men"). Default is dist1}

\item{obsName}{An optional name for the state variables being plotted (e.g., "heart rate"). Default is obsName.}

\item{minMax}{An optional vector with desired minimum and maximum quantiles to be used for setting the y-axis range on the plots, e.g., minMax <- c(.1, .9) would set the y-axis limits to the 10th and 90th percentiles of the observed state variables. If not provided, the default is to use the minimum and maximum observed values of the state variables.}

\item{printPlots}{An optional argument specifying whether the plots should be automatically printed. Default is TRUE.}

\item{numPlots}{An optional value controlling how many random examples of each profile are produced. Default is 5.}
}
\value{
The function returns the plots as a list.
}
\description{
Produces sets of prototypical example plots of the state variables' predicted temporal trajectories for each latent profile obtained based on the inertia-coordination parameters. The plots are produced by using the inertia-coordination parameters to predict temporal trajectories, with random noise added at each temporal step.
}
