% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertCoord.R
\name{inertCoordSysVarOutCompare}
\alias{inertCoordSysVarOutCompare}
\title{Compares models for predicting the system variable from the dynamic parameters.}
\usage{
inertCoordSysVarOutCompare(basedata, dist0name, dist1name, sysVarName)
}
\arguments{
\item{basedata}{A dataframe containing the inertia-coordination parameter estimates produced by the "indivInertCoord" function.}

\item{dist0name}{A name for the 0-level of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the 1-level of the distinguishing variable (e.g., "Men").}

\item{sysVarName}{A name for the system variable being predicted (e.g., "Satisfaction").}
}
\value{
The function returns a list including: 1) the lm objects containing the full results for each model for each level of the distinguishing variable (called "models0" and "models1"), 2) anova output for each model for each level of the distinguishing variable (called "anovas0" and "anovas1"), 3) summary output for each model for each level of the distinguishing variable (called "summaries0" and "summaries1") and 4) adjusted R^2 information for each model for each level of the distinguishing variable (called "adjustR20" and "adjustR21"). The function also displays histograms of the residuals and plots of the predicted values against observed values for each model.
}
\description{
The 3 models compared are the inertia-only, coordination-only and full inertia-coordination models. These are estimated separately for each level of the distinguishing variable. In other words, the system variable is predicted separately for each level of the distinguishing variable from both partner's inertia parameter estimates (inertia-only), or both partner's coordination parameter estimates (coordination-only), or both partner's inertia and coordination parameter estimates (full inertia-coordination).
}
