% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{linearSysVarOutCompare}
\alias{linearSysVarOutCompare}
\title{Compares slopes models for predicting the system variable from the dynamic parameters.}
\usage{
linearSysVarOutCompare(basedata, dist0name, dist1name, sysVarName)
}
\arguments{
\item{basedata}{A dataframe containing the slopes parameter estimates produced by the "indivLinear" function.}

\item{dist0name}{A name for the 0-level of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the 1-level of the distinguishing variable (e.g., "Men").}

\item{sysVarName}{A name for the system variable being predicted (e.g., "Satisfaction").}
}
\value{
The function returns a list including: 1) the lm objects containing the full results for each model for each level of the distinguishing variable (called "models0" and "models1"), 2) anova output for each model for each level of the distinguishing variable (called "anovas0" and "anovas1"), 3) summary output for each model for each level of the distinguishing variable (called "summaries0" and "summaries1") and 4) adjusted R^2 information for each model for each level of the distinguishing variable (called "adjustR20" and "adjustR21"). The function also displays histograms of the residuals and plots of the predicted values against observed values for each model.
}
\description{
The dynamic parameters used in these models come from a set including the linear slope for each person (linear0 and linear1) and the absolute difference between the partner's last observed values (endDif). The 3 models compared are: 1) main effects only (linear0 + linear1 + endDif), 2) the 2-way interaction of slopes, controlling for endDif (linear0*linear1 + endDif), and 3) the 3-way interaction (linear0*linear1*endDif). These are estimated separately for each level of the distinguishing variable.
}
