% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slopes.R
\name{linearSysVarOutPlots}
\alias{linearSysVarOutPlots}
\title{Plots of the system variable predicted from the slopes parameter estimates.}
\usage{
linearSysVarOutPlots(basedata, centEnd, centLine0, centLine1, sysVarName, dist,
  dist0name, dist1name)
}
\arguments{
\item{basedata}{A dataframe containing the slopes parameter estimates produced by the "indivLinear" function.}

\item{centEnd}{A vector of low, medium and high centering values for the endDif parameter estimates (e.g., values that indicate typical low, medium and high values of endDif parameter).}

\item{centLine0}{A vector of low, medium and high centering values for the slope parameter estimates for the 0-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of the slope over time for the 0-level partner).}

\item{centLine1}{A vector of low, medium and high centering values for the slope parameter estimates for the 1-level of the distinguishing variable (e.g., values that indicate typical low, medium and high values of the slope over time for the 1-level partner).}

\item{sysVarName}{A name for the system variable being predicted (e.g., "Satisfaction").}

\item{dist}{A number indicating which level of the distinguishing variable to produce plots for. Must be either 0 or 1.}

\item{dist0name}{A name for the 0-level of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the 1-level of the distinguishing variable (e.g., "Men").}
}
\description{
Displays bar or line plots for one level of the distinguishing variable, showing model predicted means and standard errors of the system variable for each of the 3 slopes models, which are: 1) at low and high levels of each of the parameter estimates (main effects), 2) at low and high levels of the two slopes, with endDif at an average level (2-way interaction), and 3) at low and high levels of all the parameters (3-way interaction).
}
