% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biVar.R
\name{makeCrossCorBiVar}
\alias{makeCrossCorBiVar}
\title{Takes typical time-series wide-format data (e.g., multiple time-varying variables for each person in wide format) and calculates cross-correlations for two user-specified variables within a specified maximum number of lags. It returns a dataframe with the largest absolute cross-correlation and its lag added for each person (e.g., it returns either the most negative or most positive cross-correlation, whichever is larger in absolute terms -- the sign is retained).}
\usage{
makeCrossCorBiVar(basedata, personId, obs1_name, obs2_name, maxLag)
}
\arguments{
\item{basedata}{The original dataframe provided by the user that includes at least two time-series variables nested within-person}

\item{personId}{The name of the column in the dataframe that has the person-level identifier.}

\item{obs1_name}{The name of the column in the dataframe that has the first time-series variable to be stacked.}

\item{obs2_name}{The name of the column in the dataframe that has the second time-series variable to be stacked.}

\item{time_name}{The name of the column in the dataframe that indicates sequential temporal observations.}
}
\value{
A cross-sectional version of the original dataframe with maximal absolute-value cross-correlations and their lags added for each person.
}
\description{
Takes typical time-series wide-format data (e.g., multiple time-varying variables for each person in wide format) and calculates cross-correlations for two user-specified variables within a specified maximum number of lags. It returns a dataframe with the largest absolute cross-correlation and its lag added for each person (e.g., it returns either the most negative or most positive cross-correlation, whichever is larger in absolute terms -- the sign is retained).
}
