% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{makeFullData}
\alias{makeFullData}
\title{Combines profile membership data from the latent profile analysis with other data for using the profile membership to predict and be predicted by the system variable.}
\usage{
makeFullData(basedata, personId, dyadId, dist, lpaData, params)
}
\arguments{
\item{lpaData}{The object created by tidyLPA's "estimate_profiles" function when "return_orig_df = TRUE"}

\item{origData}{The original dataframe provided by the user that includes all variables needed for an rties analysis, including potential control variables, etc.}

\item{lpaParams}{The object created by tidyLPA's "estimate_profiles" function when "to_return = mclust"}

\item{whichModel}{The name of the model that is being investigated (e.g., "inertCoord" or "clo")}

\item{extraVars}{An optional dataframe of cross-sectional variables to be used for more complex models that include control variables, moderators, etc. The first column must be the person ID and the second column must be the dyad ID. See the "power_user" vignette for how to make use of these variables.}
}
\value{
A list containing 1) a dataframe that contains all variables needed for using the profiles to predict, or be predicted by, the system variable (called "profileData"), and 2)a dataframe containing the profile parameter estimates needed for plotting the predicted trajectories for each profile (called "profileParams").
}
\description{
Combines profile membership data from the latent profile analysis with other data for using the profile membership to predict and be predicted by the system variable.
}
