% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{Max_Min_CCF_Signed}
\alias{Max_Min_CCF_Signed}
\title{A helper function for makeCrossCorr}
\usage{
Max_Min_CCF_Signed(a, b)
}
\arguments{
\item{a}{First time-series used in the cross-correlation}

\item{b}{Second time-series used in the cross-correlation}
}
\value{
A list of maximum absolute value cross-correlations and the lag at which they occurred.
}
\description{
A helper function for makeCrossCorr
}
