% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{plotDataByProfile}
\alias{plotDataByProfile}
\title{Plots of de-trended observed variable over time, with dyads separated into groups based on LPA profile membership.}
\usage{
plotDataByProfile(prepData, profileData, n_profiles, dist0name = NULL,
  dist1name = NULL, obs_name = NULL)
}
\arguments{
\item{prepData}{A dataframe created by the dataPrep function.}

\item{profileData}{The "profileData" dataframe created by the makeLpaData function}

\item{dist0name}{An optional name for the level-0 of the distinguishing variable (e.g., "Women"). Default is dist0.}

\item{dist1name}{An optional name for the level-1 of the distinguishing variable (e.g., "Men"). Default is dist1.}

\item{obsName}{An optional name for the observed state variable to appear on plots (e.g., "Emotional Experience").}
}
\description{
Plots of de-trended observed variable over time, with dyads separated into groups based on LPA profile membership.
}
