% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{plotLinear}
\alias{plotLinear}
\title{Produces plots of temporal trajectories predicted by linear dyadic growth models.}
\usage{
plotLinear(basedata, dist0name, dist1name, obsName)
}
\arguments{
\item{basedata}{A dataframe produced by "dataPrep".}

\item{dist0name}{A name for the 0-level of the distinguishing variable (e.g., "Women").}

\item{dist1name}{A name for the 1-level of the distinguishing variable (e.g., "Men").}

\item{obsName}{A name for the observed variable being plotted (e.g., "Emotional Experience").}
}
\description{
Produces plots of temporal trajectories predicted by linear dyadic growth models.
}
