% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesData.R
\docType{data}
\name{rties_ExampleData_1}
\alias{rties_ExampleData_1}
\title{Data for demonstrating rties models.}
\format{An object of class \code{data.frame} with 78858 rows and 6 columns.}
\usage{
rties_ExampleData_1
}
\description{
A dataset containing a minimal set of variables for demonstrating rties analyses.
}
\keyword{datasets}
