% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesDataDemo.R
\docType{data}
\name{rties_ExampleData_Demo}
\alias{rties_ExampleData_Demo}
\title{Data for demonstrating rties models.}
\format{
An object of class \code{data.frame} with 12 rows and 6 columns.
}
\usage{
rties_ExampleData_Demo
}
\description{
A dataset containing a minimal set of variables for demonstrating rties analyses.
}
\keyword{datasets}
