% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesDataExtra.R
\docType{data}
\name{rties_ExampleData_ExtraVars}
\alias{rties_ExampleData_ExtraVars}
\title{Data for demonstrating rties models.}
\format{
An object of class \code{data.frame} with 200 rows and 6 columns.
}
\usage{
rties_ExampleData_ExtraVars
}
\description{
A dataset containing a minimal set of variables for demonstrating rties analyses.
}
\keyword{datasets}
