% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesDataExampleFull.R
\docType{data}
\name{rties_ExampleDataFull}
\alias{rties_ExampleDataFull}
\title{Data for examples in the vignettes.}
\format{
An object of class \code{data.frame} with 78858 rows and 19 columns.
}
\usage{
rties_ExampleDataFull
}
\description{
A dataset containing variables for the examples in the vignettes.
}
\keyword{datasets}
