% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesDataExampleShort.R
\docType{data}
\name{rties_ExampleDataShort}
\alias{rties_ExampleDataShort}
\title{Data for the function examples.}
\format{
An object of class \code{data.frame} with 17884 rows and 19 columns.
}
\usage{
rties_ExampleDataShort
}
\description{
A dataset containing variables for the examples of the functions.
}
\keyword{datasets}
