% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{smoothData}
\alias{smoothData}
\title{Smooth one column of a dataframe with a user specified smoothing window}
\usage{
smoothData(basedata, window, personId, obs_name)
}
\arguments{
\item{basedata}{A user provided dataframe.}

\item{window}{A number specifying the window size to be smoothed over.}

\item{personId}{The variable in the dataframe specifying person ID.}

\item{obs_name}{The name of the column in the dataframe that has the time-varying observable (e.g., the variable that will be smoothed).}
}
\value{
The original dataframe with an additional column holding the smoothed observation variable, called "obsSmooth"
}
\description{
Smooth one column of a dataframe with a user specified smoothing window
}
