% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{sysVarByParam}
\alias{sysVarByParam}
\title{Produces plots of the system variable as predicted by each dyad's parameters from a given model. E.g., for the inertia-coordination model the system variable is predicted from both partner's inertia and coordination estimates.}
\usage{
sysVarByParam(paramData, colToPlot, sysVarName)
}
\arguments{
\item{paramData}{A dataframe containing parameter estimates produced by another rties function.}

\item{colToPlot}{A vector with the column numbers in the paramData for the parameter estimates that you want to be included.}

\item{sysVarName}{A name for the system variable being predicted from the parameters (e.g., "Health Behaviors").}
}
\description{
Produces plots of the system variable as predicted by each dyad's parameters from a given model. E.g., for the inertia-coordination model the system variable is predicted from both partner's inertia and coordination estimates.
}
