% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysVarInOut.R
\name{sysVarIn}
\alias{sysVarIn}
\title{Provides results for predicting couples' latent profile membership from the system variable.}
\usage{
sysVarIn(basedata, sysVar, sysVarType, n_profiles, dist0name = NULL,
  dist1name = NULL, sysVarName = NULL, printPlots = T)
}
\arguments{
\item{basedata}{A dataframe created by the makeLpaData function.}

\item{sysVar}{The name of the variable in the dataframe that contains the system variable to be predicted by profile membership.}

\item{sysVarType}{Whether the system variable is "dyadic", which means both partners have the same score, or "indiv" which means the partners can have different scores}

\item{n_profiles}{The number of latent profiles.}

\item{dist0name}{An optional name for the level-0 of the distinguishing variable (e.g., "Women"). Default is dist0.}

\item{dist1name}{An optional name for the level-1 of the distinguishing variable (e.g., "Men"). Default is dist1}

\item{sysVarName}{An optional name for the system variable being predicted (e.g., "Satisfaction"). Default is sysVar.}

\item{printPlots}{Controls whether or not the plots are printed. Default is "true".}
}
\value{
A list of model results and, by default, plots of profile membership against the system variable(s), but these can be turned off by setting printPlots=F.
}
\description{
If there are 2 profiles, then binomial regression models are used. If there are more than 2 profiles then multinomial regression is used. The system variable can be either dyadic (sysVarType = "dyadic"), where both partners have the same score (e.g., relationship length) or individual (sysVarType = "indiv"), where the partners can have different scores (e.g., age). For dyadic system variables, a couple's shared score is the only predictor of their profile membership (called "sysVar"). For individual system variables, two models are tested, one with the main effects of both partner's system variable ("sysVarMain") and one with the main effects and their interaction ("sysVarInteract"). In both cases an intercept-only model is included as a comparison point (called "base"). The function returns a list of the full model results and by default produces plots of profile membership against the system variable(s), but these can be turned off by setting printPlots=F.
}
