% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysVarInOut.R
\name{sysVarInResults}
\alias{sysVarInResults}
\title{Produces results from sysVarIn.}
\usage{
sysVarInResults(baseModel, testModel, n_profiles)
}
\arguments{
\item{baseModel}{The name of the model that was produced by sysVarIn to be used as the null model for comparison (e.g., sysIn$models$base).}

\item{testModel}{The name of the model that was produced by sysVarIn that you want results for (e.g., sysIn$models$sysVarMain or sysIn$models$sysVarInteract).}

\item{n_profiles}{The number of latent profiles.}
}
\value{
A list of results including a comparison of the test model to the null (either a LRT or Chisquare test depending on the model), a summary of the parameter estimates, exponentiated parameter estimates (e.g., odds ratios), and p values for the parameter estimates.
}
\description{
Produces results from sysVarIn.
}
