% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sysVarInOut.R
\name{sysVarOut}
\alias{sysVarOut}
\title{Provides results for predicting the system variable from the latent profiles of the dynamic parameters.}
\usage{
sysVarOut(basedata, sysVar = NULL, sysVarType, dist0name = NULL,
  dist1name = NULL, sysVarName = NULL, minMax = NULL,
  family = NULL, printPlots = T)
}
\arguments{
\item{basedata}{A dataframe created by the makeLpaData function.}

\item{sysVar}{The name of the variable in the dataframe that contains the system variable to be predicted by profile membership. If the system variable was included during the dataPrep step, then this argument should not be included. But if the system variable was only included by using the "extraVars" argument during the makeLpaData step, then this argument is required.}

\item{sysVarType}{Whether the system variable is "dyadic", which means both partners have the same score, or "indiv" which means the partners can have different scores}

\item{dist0name}{An optional name for the level-0 of the distinguishing variable to appear as plot labels (e.g., "Women"). Default is dist0.}

\item{dist1name}{An optional name for the level-1 of the distinguishing variable to appear as plot labels (e.g., "Men"). Default is dist1}

\item{sysVarName}{An optional name for the system variable to appear as plot labels (e.g., "Satisfaction"). Default is sysVar.}

\item{minMax}{An optional vector with desired minimum and maximum quantiles to be used for setting the y-axis range on the plots, e.g., minMax <- c(.1, .9) would set the y-axis limits to the 10th and 90th percentiles of the observed state variables. If not provided, the default is to use the minimum and maximum observed values of the state variables.}

\item{family}{An optional argument specifying the error distribution and link function to be used in the model. Any of the "family" options supported by glm (for dyadic system variables) or glmmPQL (for individual system variables) are available. Default is gaussian.}

\item{printPlots}{Controls whether or not the plots are printed. Default is "true".}
}
\value{
For normally distributed system variables, the function returns a list including the lm or lme objects containing the full results for each model (called "models"). Similarly, for non-normal system variables, the function returns a list of the glm or glmmPQL objects containing the full results for the models. By default, the function also displays histograms of the residuals and plots of the predicted values against observed values for each model, but these can be turned off by setting printPlots=F.
}
\description{
The system variable can be either dyadic (sysVarType = "dyadic"), where both partners have the same score (e.g., relationship length) or individual (sysVarType = "indiv"), where the partners can have different scores (e.g., age). For dyadic system variables, the only predictor is profile membership and the model is a regular regression model since all variables are at the level of the dyad. If the system variable is individual then the model is a random-intercept dyadic model and 3 models are estimated: 1) the main effect of profile membership, 2) main effects of profile membership and the distinguishing variable, and 3) the interaction of profile membership and the distinguishing variable. If the system variable is not normally distributed, any of the generalized linear models supported by glm (for dyadic system variables) or glmmPQL (for individual system variables) are available by specifying the "family" distribution.
}
