% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clo.R
\name{cloCoupledOde}
\alias{cloCoupledOde}
\title{Provides the equation for a coupled oscillator model for the differential equation solver (ode) to plot}
\usage{
cloCoupledOde(t, state, parameters)
}
\arguments{
\item{t}{A parameter used by the ode and passed by functions calling cloCoupleOde}

\item{state}{Another parameter used by the ode and passed by functions calling cloCoupleOde}

\item{parameters}{Another parameter used by the ode and passed by functions calling cloCoupleOde}
}
\value{
A list with the rates of change for each state variable.
}
\description{
Provides the equation for a coupled oscillator model for the differential equation solver (ode) to plot
}
