% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clo.R
\name{cloPredTraj}
\alias{cloPredTraj}
\title{Plots the bivariate state variable's clo model-predicted temporal trajectories for each latent profile of clo parameters.}
\usage{
cloPredTraj(origData, lpaData, lpaParams, n_profiles, time_length = NULL,
  dist0name = NULL, dist1name = NULL, obsName = NULL,
  minMax = NULL, printPlots = T)
}
\arguments{
\item{origData}{A dataframe that was produced with the "dataPrep" function.}

\item{lpaData}{A dataframe containing the LPA profile memberships.}

\item{lpaParams}{A matrix containing the clo parameter estimates associated with each of the latent profiles.}

\item{n_profiles}{The number of latent profiles.}

\item{time_length}{An optional value specifying how many time points to plot across. Default is the 75th percentile for the observed time variable.}

\item{dist0name}{An optional name for the level-0 of the distinguishing variable (e.g., "Women"). Default is dist0.}

\item{dist1name}{An optional name for the level-1 of the distinguishing variable (e.g., "Men"). Default is dist1}

\item{obsName}{An optional name for the state variables being plotted (e.g., "heart rate"). Default is obsName.}

\item{minMax}{An optional vector with desired minimum and maximum quantiles to be used for setting the y-axis range on the plots, e.g., minMax <- c(.1, .9) would set the y-axis limits to the 10th and 90th percentiles of the observed state variables. If not provided, the default is to use the minimum and maximum observed values of the state variables.}

\item{printPlots}{An optional argument specifying whether the plots should be automatically printed. Default is TRUE.}
}
\value{
The function returns the plots as a list.
}
\description{
Plots the bivariate state variable's clo model-predicted temporal trajectories for each latent profile of clo parameters.
}
