% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clo.R
\name{cloResids}
\alias{cloResids}
\title{Produces histograms of the residuals from the oscillator model for each dyad.}
\usage{
cloResids(derivData, whichModel, printPlots = T)
}
\arguments{
\item{derivData}{A dataframe that was produced with the "estDerivs" function.}

\item{whichModel}{Whether the model to be estimated is the uncoupled-oscillator ("uncoupled") or the coupled-oscillator ("coupled").}

\item{printPlots}{If true (the default) plots are displayed on the screen.}
}
\value{
A list with the histograms of the residuals for each dyad.
}
\description{
Produces histograms of the residuals from the oscillator model for each dyad.
}
\examples{
# See vignettes for examples.

}
