% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clo.R
\name{cloUncoupledOde}
\alias{cloUncoupledOde}
\title{Provides the equation for an un-coupled oscillator model for the differential equation solver (ode) to plot}
\usage{
cloUncoupledOde(t, state, parameters)
}
\description{
Provides the equation for an un-coupled oscillator model for the differential equation solver (ode) to plot
}
