% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertCoord.R
\name{crossCorPlots}
\alias{crossCorPlots}
\title{Produces cross-correlation plots of the observed state variable for lags of -+ 20 time steps for each dyad.}
\usage{
crossCorPlots(basedata, id, dyad, obs, time_name)
}
\arguments{
\item{basedata}{A user provided dataframe.}

\item{id}{The name of the column in the dataframe that has the person-level identifier.}

\item{dyad}{The name of the column in the dataframe that has the dyad-level identifier.}

\item{obs}{The name of the column in the dataframe that has the time-varying observable (e.g., the variable for which dynamics will be assessed).}

\item{time_name}{The name of the column in the dataframe that indicates sequential temporal observations.}
}
\description{
Produces cross-correlation plots of the observed state variable for lags of -+ 20 time steps for each dyad.
}
