% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inertCoord.R
\name{inertCoordResids}
\alias{inertCoordResids}
\title{Produces histograms of the residuals from the inertia-coordination model for each dyad.}
\usage{
inertCoordResids(basedata, whichModel)
}
\arguments{
\item{basedata}{A dataframe that was produced with the "dataPrep" function.}

\item{whichModel}{Whether the model to be estimated is the inertia only model ("inert"), the coordination only model ("coord"), or the full inertia-coordination model ("inertCoord").}
}
\value{
The function returns histograms of the residuals from the model for each dyad (called "plots"). The plots are also written to the working directory as a pdf file called "inertResid.pdf", or "coordResid.pdf" or "inertCoordResid.pdf"
}
\description{
Produces histograms of the residuals from the inertia-coordination model for each dyad.
}
