% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{makeBiVarData}
\alias{makeBiVarData}
\title{Takes typical long-format data, with time nested in person, and stacks two user-chosen observed variables on top of each other so they can be treated as "bivariate" within person. In other words, two time-series variables from each person are stacked on top of each other, forming a bivariate pair of variables within person. This data can then be used with the rties functions, but instead of "dyad" being the highest nesting variable, "person" is and should be substituted instead of dyad wherever needed.}
\usage{
makeBiVarData(
  basedata,
  personId,
  time_name,
  obs1_name,
  obs2_name,
  labels,
  idConvention
)
}
\arguments{
\item{basedata}{The original dataframe provided by the user that includes at least two time-series variables nested within-person}

\item{personId}{The name of the column in the dataframe that has the person-level identifier.}

\item{time_name}{The name of the column in the dataframe that indicates sequential temporal observations.}

\item{obs1_name}{The name of the column in the dataframe that has the first time-series variable to be stacked.
#' @param obs2_name The name of the column in the dataframe that has the second time-series variable to be stacked.}

\item{labels}{A string vector with the names of the variables that are being stacked.}

\item{idConvention}{The value that was added to the dist1 personID number to get the dist2 personID number}
}
\value{
A dataframe that contains the original data, plus the following columns: 1) var: the names of the stacked variables (taken from "labels"). 2) obs: the stacked observed variables, 3) dist: a zero/one distinguishing variable, and 4) varId: a variable ID that is similar to personId for use with rties. The varId for the first stacked variable is the same as the personId and the varId for the second stacked variable is personId + idConvention.
}
\description{
Takes typical long-format data, with time nested in person, and stacks two user-chosen observed variables on top of each other so they can be treated as "bivariate" within person. In other words, two time-series variables from each person are stacked on top of each other, forming a bivariate pair of variables within person. This data can then be used with the rties functions, but instead of "dyad" being the highest nesting variable, "person" is and should be substituted instead of dyad wherever needed.
}
