% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{plotDataByProfile}
\alias{plotDataByProfile}
\title{Plots of de-trended observed variable over time, with dyads separated into groups based on LPA profile membership.}
\usage{
plotDataByProfile(
  prepData,
  fullData,
  n_profiles,
  dist0name = NULL,
  dist1name = NULL,
  plot_obs_name = NULL,
  printPlots = T
)
}
\arguments{
\item{prepData}{A dataframe created by the dataPrep function.}

\item{fullData}{A dataframe created by the makeFullData function.}

\item{n_profiles}{The number of profiles that were estimated.}

\item{dist0name}{An optional name for the level-0 of the distinguishing variable (e.g., "Women"). Default is dist0.}

\item{dist1name}{An optional name for the level-1 of the distinguishing variable (e.g., "Men"). Default is dist1.}

\item{plot_obs_name}{An optional name for the observed state variable to appear on plots (e.g., "Emotional Experience").}

\item{printPlots}{If true (the default) plots are displayed on the screen.}
}
\value{
A list of plots.
}
\description{
Plots of de-trended observed variable over time, with dyads separated into groups based on LPA profile membership.
}
\examples{
# See vignettes for examples.
}
