% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{removeDyads}
\alias{removeDyads}
\title{Remove data for specified dyads from a dataframe}
\usage{
removeDyads(basedata, dyads, dyadID)
}
\arguments{
\item{basedata}{A dataframe produced by "dataPrep"}

\item{dyads}{A vector of dyad IDs to remove.}

\item{dyadID}{The variable in the dataframe specifying dyad ID; should be in the form dataframe_name$variable_name (e.g., data$couple).}
}
\value{
A dataframe with the data for the specified dyads removed.
}
\description{
Useful for cleaning data if some dyads have extensive missing or otherwise problematic data.
}
