% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesShared.R
\name{removeDyads}
\alias{removeDyads}
\title{Remove data for specified dyads from a dataframe}
\usage{
removeDyads(basedata, dyads, dyadId)
}
\arguments{
\item{basedata}{A user provided dataframe.}

\item{dyads}{A vector of dyad IDs to remove.}

\item{dyadId}{The variable in the dataframe specifying dyad ID.}
}
\value{
A dataframe with the data for the specified dyads removed.
}
\description{
Useful for cleaning data if some dyads have extensive missing or otherwise problematic data.
}
\examples{
data <- rties_ExampleDataShort
dyads <- c(3, 12)
newData <- removeDyads(basedata=data, dyads=dyads, dyadId="couple")

}
