% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtiesData2.R
\docType{data}
\name{rties_ExampleData_2}
\alias{rties_ExampleData_2}
\title{Data for demonstrating rties models.}
\format{An object of class \code{data.frame} with 31308 rows and 7 columns.}
\usage{
rties_ExampleData_2
}
\description{
A dataset containing a minimal set of variables for demonstrating rties analyses.
}
\keyword{datasets}
